WIP - adding standalone for loop block
